feat: get weekly
일자별 요약 7일치 구현 완료

+++ src/main/java/kr/casealot/shop/domain/function/dto/FunctionDTO.java
@@ -13,9 +13,9 @@ import lombok.Setter;
 @AllArgsConstructor
 public class FunctionDTO {
 
-  private int todayOrder;
-  private int todayCancel;
-  private int todayReturn;
-  private int todayChange;
-  private int todayQna;
+  private long todayOrder;
+  private long todayCancel;
+  private long todayReturn;
+  private long todayChange;
+  private long todayQna;
 }
